% Generated by roxygen2 (4.0.2): do not edit by hand
\name{movhm.diff}
\alias{movhm.diff}
\title{Create a difference heatmap based on movement data}
\usage{
movhm.diff(lx, ly, x, y, difference = "relative", blocksize, margins,
  origin = NULL, consider.time = FALSE, time.transformation = sqrt,
  print = FALSE)
}
\arguments{
\item{lx}{a list of data frames containing columns for x- and y-values}

\item{ly}{a list of data frames containing columns for x- and y-values}

\item{x}{name of the column containing x-values in the data frames in
\code{lx} and \code{ly} (string)}

\item{y}{name of the column containing y-values in the data frames in
\code{lx} and \code{ly} (string)}

\item{difference}{a character string specifying whether to compute
\code{"relative"} (default) or \code{"absolute"} differences.}

\item{blocksize}{scaling factor}

\item{margins}{numeric vector of length 4 containing the margins (xmin, ymin,
xmax, ymax)}

\item{origin}{(optional) numeric vector of length 2 (x and y) of the position
to be removed from the heatmap (e.g. starting position)}

\item{consider.time}{logical indicating whether to count one subject multiple
times in one cell, depending on the time the subject was inside the cell.}

\item{time.transformation}{if \code{consider.time} is \code{TRUE}, then
\code{time.transformation} specifies the function to be applied to the
position * time frequencies (default is \code{sqrt}). Can also be
\code{NULL}.}

\item{print}{logical indicating if output should be printed via ggplot}
}
\value{
A data frame with x, y and f (freqency) columns ready for plotting
  with ggplot2
}
\description{
Create a difference heatmap based on movement data
}
\examples{
library(dplyr)
library(ggplot2)

data(movdat)

movhm.diff(movdat[1:21], movdat[22:42], x = "cart_x", y = "cart_y",
           blocksize = 50,  margins = c(-750, -2000, 100, 1550),
           origin = c(-120, -1000)) \%>\%
  ggplot(aes(x, y, fill = f)) +
  geom_raster() +
  scale_fill_gradientn(colours = scale_RdBu, na.value = rgb(0, 0, 0, 0),
                       limits = c(-.4, .4)) +
  coord_fixed() +
  theme_movhm()
}

