% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movhm.R
\name{movhm}
\alias{movhm}
\title{Two-dimensional binning for group movement data}
\usage{
movhm(l, x, y, bin_width, margins = NULL, drop = TRUE, time = TRUE,
  time_transform = NULL, plot = FALSE)
}
\arguments{
\item{l}{A list of data frames or named lists containing x- and y-values}

\item{x}{character string indicating the name of the variable holding
x-values in the data frames or lists in \code{l}.}

\item{y}{character string indicating the name of the variable holding
y-values in the data frames or lists in \code{l}.}

\item{bin_width}{numeric indicating the size of the bins}

\item{margins}{numeric vector of length 4 indicating the margins (xmin, ymin,
xmax, ymax)}

\item{drop}{logical indicating whether to set empty bins to NA instead of 0
(default: \code{TRUE}). This is useful when plotting, so empty bins can be
plotted transparent.}

\item{time}{logical indicating whether to count one case multiple
times in one cell, depending on the time the subject was inside the cell.}

\item{time_transform}{if \code{time} is \code{TRUE}, then
\code{time_transform} specifies the function to be applied to the
position * time frequencies, e.g. \code{sqrt}.}

\item{plot}{logical indicating if the heatmap should be plotted}
}
\value{
A data frame with x, y and f (freqency) columns
}
\description{
Two-dimensional binning for group movement data
}
\examples{
data(movdat)

bins <- movhm(movdat, x = "cart_x", y = "cart_y", bin_width = 50,
              margins = c(-750, -2000, 100, 1550))

library(ggplot2)
ggplot(bins, aes(x, y, fill = f)) +
  geom_raster() +
  scale_fill_gradientn(colours = scale_Blues, na.value = rgb(0, 0, 0, 0)) +
  coord_fixed() +
  theme_movhm()

}

