% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movhm.R
\name{movhm_diff}
\alias{movhm_diff}
\title{movhm_diff}
\usage{
movhm_diff(lx, ly, x, y, bin_width, difference = c("relative", "absolute"),
  margins = NULL, time = TRUE, time_transform = NULL, plot = FALSE)
}
\arguments{
\item{lx}{A list of data frames or named lists containing x- and y-values}

\item{ly}{A list of data frames or named lists containing x- and y-values}

\item{x}{character string indicating the name of the variable holding
x-values in the data frames or lists in \code{l}.}

\item{y}{character string indicating the name of the variable holding
y-values in the data frames or lists in \code{l}.}

\item{bin_width}{numeric indicating the size of the bins}

\item{difference}{character string indicating whether absolute
(\code{"absolute"}) or relative (\code{"relative"}) frequency differences
should be calculated.}

\item{margins}{numeric vector of length 4 indicating the margins (xmin, ymin,
xmax, ymax)}

\item{time}{logical indicating whether to count one case multiple
times in one cell, depending on the time the subject was inside the cell.}

\item{time_transform}{if \code{time} is \code{TRUE}, then
\code{time_transform} specifies the function to be applied to the
position * time frequencies, e.g. \code{sqrt}.}

\item{plot}{logical indicating if the heatmap should be plotted}
}
\description{
movhm_diff
}

