% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{movhmcolgradients}
\alias{movhmcolgradients}
\alias{scale_Blues}
\alias{scale_RdBu}
\alias{scale_RdGrBu}
\alias{scale_Spectral}
\title{Colour gradients}
\format{\preformatted{ chr [1:100] "#1639FA" "#183EFA" "#1B43FA" "#1D49FA" "#204EFA" ...
}}
\usage{
scale_RdGrBu

scale_Blues

scale_RdBu

scale_Spectral
}
\description{
These colour gradients are used in conjunction with ggplot2's
\code{scale_fill_gradientn()}. Available gradients are

\describe{
  \item{\code{scale_Blues}}{
  A colour gradient for heatmaps, from white to blue
  }
  \item{\code{scale_RdGrBu}}{
  A colour gradient for difference heatmaps, from red over green to blue
  }
  \item{\code{scale_RdBu}}{
  A colour gradient for difference heatmaps, from red over white to blue
  }
  \item{\code{scale_Spectral}}{
  A colour gradient for difference heatmaps, from red over yellow to blue
  }
}
}
\seealso{
\code{\link{scale_fill_gradientn}}
}
\keyword{datasets}

