% Generated by roxygen2 (4.0.1): do not edit by hand
\name{roi.analysis}
\alias{roi.analysis}
\title{Execute a ROI analysis on movement data}
\usage{
roi.analysis(..., roi, blocksize, margins)
}
\arguments{
\item{...}{lists of two-column data frames with x- and y-values}

\item{roi}{a two-column data frame with x- and y-values specifying what cells
should be included in the ROI.}

\item{blocksize}{scaling factor}

\item{margins}{numeric vector of length 4 containing the margins (xmin, ymin,
xmax, ymax)}
}
\description{
Execute a ROI analysis on movement data
}

