% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roi_analysis.R
\docType{class}
\name{ROI}
\alias{ROI}
\title{ROI Class}
\format{A \code{\link{ROI}} generator object}
\usage{
\code{ROI$new(bin_width)}
}
\description{
Class for defining a ROI (area of interest) for use with \code{roi_analysis}.
}
\section{Methods}{

\describe{
  \item{\code{add(xmin, ymin, xmax, ymax)}}{Add a region to the ROI.}
  \item{\code{plot(bg = NULL)}}{Plot the ROI (to an optional background).}
}
}
\examples{
# Create a new roi with a bin width of 1
roi <- ROI$new(1)

# Add the area from 0/0 to 2/4 to the ROI and view it
roi$add(0, 0, 2, 4)
roi$plot()

# Increase the roi
roi$add(2, 2, 4, 4)
roi$plot()
}
\keyword{datasets}

