% Generated by roxygen2 (4.0.2): do not edit by hand
\name{show.me.da.raster}
\alias{show.me.da.raster}
\title{Display the grid for roi analysis}
\usage{
show.me.da.raster(template, blocksize, margins, roi = NULL)
}
\arguments{
\item{template}{a list of ggplot2-geoms which will be plotted in the
background.}

\item{blocksize}{scaling factor}

\item{margins}{numeric vector of length 4 containing the margins (xmin, ymin,
xmax, ymax)}

\item{roi}{a two-column data frame with x- and y-values specifying what cells
should be included in the ROI.}
}
\description{
Display the grid for roi analysis
}

