% Generated by roxygen2 (4.0.2): do not edit by hand
\name{theme_movhm}
\alias{theme_movhm}
\title{A blank theme for heatmaps}
\usage{
theme_movhm(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
A blank theme for heatmaps
}

