% Generated by roxygen2 (4.0.1): do not edit by hand
\name{count.pos}
\alias{count.pos}
\title{Count occurrences in position raster}
\usage{
count.pos(x, blocksize, margins, consider.time = FALSE)
}
\arguments{
\item{x}{a data frame containing columns for x- and y-values}

\item{blocksize}{scaling factor}

\item{margins}{numeric vector of length 4 containing the margins (xmin, ymin,
xmax, ymax)}

\item{consider.time}{logical indicating whether to count one subject multiple
times in one cell}
}
\value{
A data frame with x, y and f (freqency) columns
}
\description{
Count occurrences in position raster
}

