% Generated by roxygen2 (4.0.1): do not edit by hand
\name{movhm.diff}
\alias{movhm.diff}
\title{Create a difference heatmap based on movement data}
\usage{
movhm.diff(lx, ly, difference = "relative", blocksize, margins,
  origin = NULL, consider.time = FALSE, print = FALSE)
}
\arguments{
\item{lx}{a list of data frames containing columns for x- and y-values}

\item{ly}{a list of data frames containing columns for x- and y-values}

\item{difference}{a character string specifying whether to compute
\code{"relative"} (default) or \code{"absolute"} differences.}

\item{blocksize}{scaling factor}

\item{margins}{numeric vector of length 4 containing the margins (xmin, ymin,
xmax, ymax)}

\item{origin}{(optional) numeric vector of length 2 (x and y) of the position
to be removed from the heatmap (e.g. starting position)}

\item{consider.time}{logical indicating whether to count one subject multiple
times in one cell}

\item{print}{logical indicating if output should be printed via ggplot}
}
\value{
A data frame with x, y and f (freqency) columns ready for plotting
  with ggplot2
}
\description{
Create a difference heatmap based on movement data
}
\examples{
\dontrun{
movhm(mylist.a, mylist.b, blocksize = 50, margins = c(0, 0, 2000, 300),
      origin = c(1200, 100)) \%>\%
  ggplot(aes(x, y, fill = f)) +
  geom_raster() +
  scale_fill_gradientn(colours = scale_RdBu, na.value = rgb(0, 0, 0, 0)) +
  coord_fixed() +
  theme_movhm()
}
}

