% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gradients.R
\docType{data}
\name{movhmcolgradients}
\alias{movhmcolgradients}
\alias{scale_Blues}
\alias{scale_RdBu}
\alias{scale_RdBu_bias}
\alias{scale_RdGrBu}
\alias{scale_Spectral}
\alias{scale_Spectral_bias}
\title{Colour gradients}
\format{\preformatted{ chr [1:100] "#FF0000" "#FF0A00" "#FF1400" "#FF1E00" "#FF2800" ...
}}
\usage{
scale_RdGrBu

scale_Blues

scale_RdBu

scale_RdBu_bias(bias = 1.5)

scale_Spectral

scale_Spectral_bias(bias = 1.5)
}
\description{
These colour gradients are used in conjunction with ggplot2's
\code{scale_fill_gradientn()}. Available gradients are

\describe{
  \item{\code{scale_Blues}}{
  A colour gradient for heatmaps, from white to blue
  }
  \item{\code{scale_RdGrBu}}{
  A colour gradient for difference heatmaps, from red over green to blue
  }
  \item{\code{scale_RdBu}}{
  A colour gradient for difference heatmaps, from red over white to blue
  }
  \item{\code{scale_RdBu_bias(bias = 1.5)}}{
  A function for creating a colour gradient for difference heatmaps, from red
  over white to blue (like scale_RdBu). The \code{bias} parameter defines
  how widely the colours space towards both ends.
  }
  \item{\code{scale_Spectral}}{
  A colour gradient for difference heatmaps, from red over yellow to blue
  }
  \item{\code{scale_Spectral_bias(bias = 1.5)}}{
  A function for creating a colour gradient for difference heatmaps, from red
  over yellow to blue (like scale_Spectral). The \code{bias} parameter
  defines how widely the colours space towards both ends.
  }
}
}
\seealso{
\code{\link{scale_fill_gradientn}}
}
\keyword{datasets}

