% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roi_analysis.R
\docType{class}
\name{ROI}
\alias{ROI}
\title{ROI Class}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x35fb558> 
 - attr(*, "name")= chr "ROI_generator"
}}
\usage{
ROI
}
\description{
ROI Class
}
\examples{
# Create a new roi with a bin width of 1
roi <- ROI$new(1)

# Add the area from 0/0 to 2/4 to the ROI and view it
roi$add(0, 0, 2, 4)
roi$plot()

# Increase the roi
roi$add(2, 2, 4, 4)
roi$plot()
}
\keyword{datasets}

